Moved all scripts into the dedicated folder
