+++ src/pages/index.html
@@ -152,6 +152,6 @@
 		<div id="popup-backdrop"></div>
 
 		<!-- js -->
-		<script type="module" src="/merge/script.js"></script>
+		<script type="module" src="/merge/index.js"></script>
 	</body>
 </html>
