+++ src/scripts/state.ts
